+++ organic_chem_hub/urls.py
@@ -1,12 +1,18 @@
 from django.contrib import admin
 from django.conf import settings
 from django.conf.urls.static import static
+from django.contrib.sitemaps.views import sitemap
 from django.urls import include, path
+
+from reactions.sitemaps import sitemaps
+from reactions.views import robots_txt_view
 
 
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("accounts/", include("django.contrib.auth.urls")),
+    path("sitemap.xml", sitemap, {"sitemaps": sitemaps}, name="django.contrib.sitemaps.views.sitemap"),
+    path("robots.txt", robots_txt_view, name="robots_txt"),
     path("", include("reactions.urls")),
 ]